feat: implement pairing heaps
This includes an alternative attempt at making a version with borrowing.
However this does not work there an idea to have the nodes be owned by
the lookup table but since the expected gain is not going to be that
great we chose the eary way of letting the nodes have multiple owners
using Rc.